move memry util to io package
