+++ java/runtime/src/main/java/io/ray/runtime/io/BufferInputStream.java
@@ -1,6 +1,5 @@
 package io.ray.runtime.io;
 
-import io.ray.runtime.util.MemoryBuffer;
 import java.io.IOException;
 import java.io.InputStream;
