+++ java/runtime/src/main/java/io/ray/runtime/io/BufferWritableByteChannel.java
@@ -1,6 +1,5 @@
 package io.ray.runtime.io;
 
-import io.ray.runtime.util.MemoryBuffer;
 import java.nio.ByteBuffer;
 import java.nio.channels.WritableByteChannel;
