+++ java/runtime/src/main/java/io/ray/runtime/io/MemoryBuffer.java
@@ -1,4 +1,4 @@
-package io.ray.runtime.util;
+package io.ray.runtime.io;
 
 import static com.google.common.base.Preconditions.checkArgument;
