+++ java/runtime/src/main/java/io/ray/runtime/io/MemoryUtils.java
@@ -1,4 +1,4 @@
-package io.ray.runtime.util;
+package io.ray.runtime.io;
 
 import java.nio.ByteBuffer;
