+++ java/runtime/src/main/java/io/ray/runtime/io/Platform.java
@@ -1,4 +1,4 @@
-package io.ray.runtime.util;
+package io.ray.runtime.io;
 
 import com.google.common.base.Preconditions;
 import java.lang.reflect.Constructor;
@@ -343,6 +343,8 @@ public final class Platform {
     return offHeapAddress;
   }
 
+  private static final ByteBuffer localBuffer = ByteBuffer.allocateDirect(0);
+
   /** Wrap a buffer [address, address + size) into provided <code>buffer</code>. */
   public static void wrapDirectBuffer(ByteBuffer buffer, long address, int size) {
     UNSAFE.putLong(buffer, BUFFER_ADDRESS_FIELD_OFFSET, address);
@@ -352,7 +354,7 @@ public final class Platform {
 
   public static ByteBuffer wrapDirectBuffer(long address, int size) {
     try {
-      ByteBuffer buffer = (ByteBuffer) UNSAFE.allocateInstance(DIRECT_BYTE_BUFFER_CLASS);
+      ByteBuffer buffer = localBuffer.duplicate();
       UNSAFE.putLong(buffer, BUFFER_ADDRESS_FIELD_OFFSET, address);
       UNSAFE.putInt(buffer, BUFFER_CAPACITY_FIELD_OFFSET, size);
       buffer.clear();
