+++ java/runtime/src/test/java/io/ray/runtime/io/MemoryBufferTest.java
@@ -1,4 +1,4 @@
-package io.ray.runtime.util;
+package io.ray.runtime.io;
 
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
